Just move mixin code to the only component using it
